allow clusterprops setting through set-circuit-breakers clusterprops API endpoint
