+++ solr/core/src/java/org/apache/solr/util/circuitbreaker/GlobalCircuitBreakerManager.java
@@ -5,13 +5,11 @@ import java.io.IOException;
 import java.lang.invoke.MethodHandles;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
-import java.util.concurrent.ConcurrentHashMap;
 import org.apache.solr.client.solrj.SolrRequest;
-import org.apache.solr.common.annotation.JsonProperty;
 import org.apache.solr.common.cloud.ClusterPropertiesListener;
 import org.apache.solr.common.util.Utils;
 import org.apache.solr.core.CoreContainer;
+import org.apache.solr.core.GlobalCircuitBreakerConfig;
 import org.apache.solr.util.SolrJacksonAnnotationInspector;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -41,52 +39,6 @@ public class GlobalCircuitBreakerManager implements ClusterPropertiesListener {
         super();
         this.factory = new GlobalCircuitBreakerFactory(coreContainer);
         this.coreContainer = coreContainer;
-    }
-
-    private static class GlobalCircuitBreakerConfig {
-        static final String CIRCUIT_BREAKER_CLUSTER_PROPS_KEY = "circuit-breakers";
-        @JsonProperty Map<String, CircuitBreakerConfig> configs = new ConcurrentHashMap<>();
-
-        @JsonProperty
-        Map<String, Map<String, CircuitBreakerConfig>> hostOverrides = new ConcurrentHashMap<>();
-
-        static class CircuitBreakerConfig {
-            @JsonProperty Boolean enabled = false;
-            @JsonProperty Boolean warnOnly = false;
-            @JsonProperty Double updateThreshold = Double.MAX_VALUE;
-            @JsonProperty Double queryThreshold = Double.MAX_VALUE;
-
-            @Override
-            public int hashCode() {
-                return Objects.hash(enabled, warnOnly, updateThreshold, queryThreshold);
-            }
-
-            @Override
-            public boolean equals(Object obj) {
-                if (obj instanceof CircuitBreakerConfig) {
-                    CircuitBreakerConfig that = (CircuitBreakerConfig) obj;
-                    return that.enabled.equals(this.enabled)
-                            && that.warnOnly.equals(this.warnOnly)
-                            && that.updateThreshold.equals(this.updateThreshold)
-                            && that.queryThreshold.equals(this.queryThreshold);
-                }
-                return false;
-            }
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (o instanceof GlobalCircuitBreakerConfig) {
-                GlobalCircuitBreakerConfig that = (GlobalCircuitBreakerConfig) o;
-                return that.configs.equals(this.configs) && that.hostOverrides.equals(this.hostOverrides);
-            }
-            return false;
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(configs, hostOverrides);
-        }
     }
 
     // for registering global circuit breakers set in clusterprops
